# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ApplyBlitz API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
    put:
      operationId: upsertProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # === JOBS ===
  /jobs/search:
    get:
      operationId: searchJobs
      tags: [jobs]
      summary: Search for jobs from external APIs
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: boolean
        - name: minSalary
          in: query
          schema:
            type: ["integer", "null"]
        - name: maxSalary
          in: query
          schema:
            type: ["integer", "null"]
        - name: datePosted
          in: query
          schema:
            type: string
            enum: [today, week, month]
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Job search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobListing"
  /jobs/saved:
    get:
      operationId: getSavedJobs
      tags: [jobs]
      summary: Get all saved jobs
      responses:
        "200":
          description: Saved jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedJob"
    post:
      operationId: saveJob
      tags: [jobs]
      summary: Save a job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedJobInput"
      responses:
        "201":
          description: Job saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedJob"
  /jobs/saved/{id}:
    delete:
      operationId: deleteSavedJob
      tags: [jobs]
      summary: Delete a saved job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === APPLICATIONS ===
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, name, targetRoles, targetLocations, yearsExperience, skills, resumeText]
      properties:
        id:
          type: integer
        name:
          type: string
        targetRoles:
          type: string
        targetLocations:
          type: string
        yearsExperience:
          type: integer
        skills:
          type: string
        resumeText:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        emailUser:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProfileInput:
      type: object
      required: [name, targetRoles, targetLocations, yearsExperience, skills, resumeText]
      properties:
        name:
          type: string
        targetRoles:
          type: string
        targetLocations:
          type: string
        yearsExperience:
          type: integer
        skills:
          type: string
        resumeText:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        emailUser:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]

    # --- Job Listings ---
    JobListing:
      type: object
      required: [externalId, title, company, location, source, applyUrl, description, postedAt]
      properties:
        externalId:
          type: string
        title:
          type: string
        company:
          type: string
        location:
          type: string
        source:
          type: string
        applyUrl:
          type: string
        description:
          type: string
        salary:
          type: ["string", "null"]
        matchScore:
          type: ["integer", "null"]
        isRemote:
          type: boolean
        postedAt:
          type: string
    SavedJob:
      type: object
      required: [id, externalId, title, company, location, source, applyUrl, description, postedAt, createdAt]
      properties:
        id:
          type: integer
        externalId:
          type: string
        title:
          type: string
        company:
          type: string
        location:
          type: string
        source:
          type: string
        applyUrl:
          type: string
        description:
          type: string
        salary:
          type: ["string", "null"]
        matchScore:
          type: ["integer", "null"]
        isRemote:
          type: boolean
        postedAt:
          type: string
        createdAt:
          type: string
    SavedJobInput:
      type: object
      required: [externalId, title, company, location, source, applyUrl, description, postedAt]
      properties:
        externalId:
          type: string
        title:
          type: string
        company:
          type: string
        location:
          type: string
        source:
          type: string
        applyUrl:
          type: string
        description:
          type: string
        salary:
          type: ["string", "null"]
        matchScore:
          type: ["integer", "null"]
        isRemote:
          type: boolean
        postedAt:
          type: string

    # --- Applications ---
...