# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ApplyBlitz API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  targetRoles: string;
  targetLocations: string;
  yearsExperience: number;
  skills: string;
  resumeText: string;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  /** @nullable */
  emailUser?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  createdAt?: string;
  updatedAt?: string;
}