# lib/api-zod/src/generated/types/savedJobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ApplyBlitz API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SavedJobInput {
  externalId: string;
  title: string;
  company: string;
  location: string;
  source: string;
  applyUrl: string;
  description: string;
  /** @nullable */
  salary?: string | null;
  /** @nullable */
  matchScore?: number | null;
  isRemote?: boolean;
  postedAt: string;
}